Externalized the opensearch exceptions in their own file.
