+++ lib/charms/opensearch/v0/opensearch_exceptions.py
@@ -1,0 +1,62 @@
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""File containing all OpenSearch related exceptions."""
+from typing import List
+
+# The unique Charmhub library identifier, never change it
+LIBID = "9e5bbab8a2bb475d83252500481351b2"
+
+# Increment this major API version when introducing breaking changes
+LIBAPI = 0
+
+# Increment this PATCH version before using `charmcraft publish-lib` or reset
+# to 0 if you are raising the major API version
+LIBPATCH = 1
+
+
+class OpenSearchError(Exception):
+    """Base exception class for OpenSearch errors."""
+
+
+class OpenSearchMissingError(OpenSearchError):
+    """Exception thrown when an action is attempted on OpenSearch when it's not installed."""
+
+
+class OpenSearchInstallError(OpenSearchError):
+    """Exception thrown when OpenSearch fails to be installed."""
+
+
+class OpenSearchMissingSysReqError(OpenSearchError):
+    """Exception thrown when OpenSearch fails to be installed."""
+
+    def __init__(self, missing_requirements: List[str]):
+        self.missing_requirements = missing_requirements
+
+
+class OpenSearchStartError(OpenSearchError):
+    """Exception thrown when OpenSearch fails to start."""
+
+
+class OpenSearchStopError(OpenSearchError):
+    """Exception thrown when OpenSearch fails to stop."""
+
+
+class OpenSearchRestartError(OpenSearchError):
+    """Exception thrown when OpenSearch fails to restart."""
+
+
+class OpenSearchNotStartedError(OpenSearchError):
+    """Exception thrown when attempting an operation when the OpenSearch service is stopped."""
+
+
+class OpenSearchCmdError(OpenSearchError):
+    """Exception thrown when an OpenSearch bin command fails."""
+
+
+class OpenSearchHttpError(OpenSearchError):
+    """Exception thrown when an OpenSearch REST call fails."""
+
+
+class OpenSearchHAError(OpenSearchError):
+    """Exception thrown when the HA of the OpenSearch charm is violated."""
