+++ src/charm.py
@@ -12,7 +12,7 @@ from charms.opensearch.v0.constants_charm import InstallError, InstallProgress
 from charms.opensearch.v0.constants_tls import CertType
 from charms.opensearch.v0.helper_security import to_pkcs8
 from charms.opensearch.v0.opensearch_base_charm import OpenSearchBaseCharm
-from charms.opensearch.v0.opensearch_distro import OpenSearchInstallError
+from charms.opensearch.v0.opensearch_exceptions import OpenSearchInstallError
 from ops.charm import InstallEvent
 from ops.main import main
 from ops.model import BlockedStatus, MaintenanceStatus
